// src/types/rastreamentoTubaroes.ts
export interface RastreamentoTubaroes {
  Id: number
  Tempo: string
  Lat: number
  Lon: number
  TempCc: number
  PForrageio: number
  Comportamento: string
  ChlorAAmbiente: number
  SshaAmbiente: number
}

export interface RastreamentoTubaroesDto {
  Id: number
  Tempo: string
  Lat: number
  Lon: number
  TempCc: number
  PForrageio: number
  Comportamento: string
  ChlorAAmbiente: number
  SshaAmbiente: number
}